requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pypdf
google-genai
python-multipart